- Updated pre-select.ks to prompt user for Master IP,Netmask,Gateway, and DNS - Deploys Traefik Ingress as part of initial build (at install time) - Moved Cockpit web-interface to ingress
NOTE: Traefik is designed to utilize an FQDN that has a corresponding DNS entry. The build system will attempt to detect the servers DNS name at installation time. If it does not have a DNS entry for the cluster, then it will use the name "edcop-master.local" as the entry. If this happen, a DNS entry must be created for "edcop-master.local" after the cluster has been installed.